Revert "Moved Drive constants from the constant file to the DriveSubsystem file"
This reverts commit 7a63adb2c1a63ba2d47dc30e82c7e5c14f686fa0.
